+++ helper/widgets/text.js
@@ -28,18 +28,28 @@ Text.constitute(function prepareSchema() {
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.1.0
- * @version  0.1.0
+ * @version  0.1.6
  *
  * @param    {HTMLElement}   widget
  */
 Text.setMethod(function populateWidget() {
 
+	let tag_name;
+
 	populateWidget.super.call(this);
 
-	let paragraph = this.createElement('p');
-	paragraph.textContent = this.config.content || '';
+	if (this.widget.dataset.textElementTag) {
+		tag_name = this.widget.dataset.textElementTag;
+	}
 
-	this.widget.append(paragraph);
+	if (!tag_name) {
+		tag_name = 'p';
+	}
+
+	let text_element = this.createElement(tag_name);
+	text_element.textContent = this.config.content || '';
+
+	this.widget.append(text_element);
 });
 
 /**